Fixed Position Nav and Header
Header and nav now have fixed position.

Done entirely with CSS so that it'll be easier to port to Tumblr.

The idea is to revert the header to default static positioning when in
overhead view or lower widths, and move the Nav portion to underneath
the breadcrumb in a static horizontal layout.

The defaults of bootstrap still seem to cause a weird margin to appear
on the left hand side when window width is decreased enough. Not sure
exactly what is causing that though, so will need to try tracking it
down.

There's quite a few modifications to both the CSS and HTML in order to
get things to consistently line up when resizing.