z-index fix & index breadcrumb change
Added a z-index to the header to ensure that it didn't end layered
behind the nav section, which was preventing the right most buttons on
the top-nav from being clickable on some browsers.
Changed the breadcrumb to indicate that the current page is "Home"
rather than "Week 2: tools and xhtml".